refactor : 상수 내용 및 import부분 수정

+++ src/main/java/lotto/PrizeProfit.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public class PrizeProfit {
     private final double rate;
 
-    public PrizeProfit(PrizeStats prizeStats, long buyingPrice) {
+    public PrizeProfit(PrizeStats prizeStats, int buyingPrice) {
         this.rate = calculateRate(calculatePrizeMoney(prizeStats), buyingPrice);
     }
 
@@ -13,8 +13,8 @@ public class PrizeProfit {
         return this.rate;
     }
 
-    private long calculatePrizeMoney(PrizeStats prizeStats) {
-        long sumOfPrizeMoney = 0;
+    private int calculatePrizeMoney(PrizeStats prizeStats) {
+        int sumOfPrizeMoney = 0;
         for (Map.Entry<PrizeGrade, Integer> entry : prizeStats.getGradeDist().entrySet()) {
             sumOfPrizeMoney += entry.getKey().getPrizeMoney() * entry.getValue();
         }
